Rename events to EventProps and place queries and mutations inside hooks folder
